+++ django_willpower/data/default_stack/tests/models.py
@@ -7,19 +7,20 @@ import {{ app.code }}.models as app_models
 import {{ app.code }}.factories as app_factories
 {% for field in model_inventory.modelfields %}{% if field.choices_list %}from {{ app.code }}.choices import get_{{ field.name }}_default
 {% endif %}{% endfor %}
-{% set default_field = model_inventory.string_representation %}
+{% set representation_field = utils.get_model_representation_string(model_inventory.string_representation) %}
+{% set default_field = model_inventory.get_required_fields() | first %}
 def test_basic(db):
     """
     Basic model validation with required fields should not fail
     """
-    obj = app_models.{{ model_inventory.name }}({% for field in model_inventory.modelfields %}{% if field.required %}
-        {{ field.name }}={{ utils.test_value(field) }},{% endif %}{% endfor %}
+    obj = app_models.{{ model_inventory.name }}({% for field in model_inventory.get_required_fields() %}
+        {{ field.name }}={{ utils.test_dummy_value(field) }},{% endfor %}
     )
     obj.full_clean()
     obj.save()
 
-    assert app_models.{{ model_inventory.name }}.objects.filter({{ default_field }}="Foo").count() == 1
-    assert obj.{{ default_field }} == "Foo"
+    assert app_models.{{ model_inventory.name }}.objects.filter({{ default_field.name }}={{ utils.test_dummy_value(default_field) }}).count() == 1
+    assert obj.{{ default_field.name }} == {{ utils.test_dummy_value(default_field) }}
     assert obj.get_absolute_url() is not None
 
 
@@ -35,15 +36,15 @@ def test_required_fields(db, settings):
     with pytest.raises(ValidationError) as excinfo:
         obj.full_clean()
 
-    assert excinfo.value.message_dict == {
-        "heurist_id": ["This field cannot be blank."],
-        "langage": ["This field cannot be blank."],
-    }
+    assert excinfo.value.message_dict == {{ '{' }}{% for field in model_inventory.get_required_fields() %}
+        "{{ field.name }}": ["This field cannot be blank."],{% endfor %}
+    {{ '}' }}
 
 
 def test_factory_creation(db):
     """
     Factory should correctly create a new object without any errors
     """
-    obj = app_factories.{{ model_inventory.name }}Factory({{ default_field }}="foo")
-    assert obj.{{ default_field }} == "foo"
+    obj = app_factories.{{ model_inventory.name }}Factory({{ representation_field }}="foo")
+    assert obj.{{ representation_field }} == "foo"
+